backend/ingestion/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
pgvector==0.2.4
pypdf==4.0.1
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
